Add more attributes to cafeteriaEntity, refactor hooks and fetches
